+++ src/styles/blocks/teachers.scss
@@ -86,7 +86,7 @@
       height: 100%;
       background-color: $controls-bg-secondary;
       border-radius: 10px;
-      transition: transform 0.5s ease;
+      transition: transform 0.3s ease;
 
       &:hover {
         transform: scale(1.03);
